server/requirements.txt
Flask>=2.2
python-dotenv
redis
